ra-spectral: cast gated corrections to the attention dtype
The Stage-4 seed-0 scalar-delta arm crashed on its first forward:
under bf16 autocast the pointwise gate modes promote to fp32 (the
fp32 beta parameter multiplies the bf16 delta) while einsum-based
spectral modes stay bf16, and the indexed head assignment requires
matching dtypes. Cast the corrected heads to the attention output
dtype at assignment, and add a bf16-autocast forward regression
test over the scalar, coordinate, and spectral modes — the gap the
CPU-fp32 tests and the fp32 oracle both missed. Seed 0 therefore
lacks its scalar-delta arm; the patched file was hot-pushed to the
running pod before seeds 1 and 2 reached that arm.

Generated-by: Claude Fable 5

+++ fim/reciprocal_attention/gpt2_spectral_ra.py
@@ -238,7 +238,11 @@ class SpectralDeltaAttention(CausalSelfAttention_KNLP):
                     "bstr,sdr->bstd", modes * beta[None, :, None, :], self.basis_u
                 )
             yb = yb.clone()
-            yb[:, sel] = std_sel + corr
+            # Under bf16 autocast the pointwise gate modes (scalar,
+            # coordinate) promote to fp32 (fp32 beta parameter times
+            # bf16 delta) while einsum-based modes stay bf16; index
+            # assignment requires matching dtypes, so cast explicitly.
+            yb[:, sel] = (std_sel + corr).to(yb.dtype)
 
         y = yb.transpose(1, 2).contiguous().view(B, T, C)
         y = self.resid_dropout(self.c_proj(y))
